courses/build_production_ready_agents/ch3_demos/requirements.txt
google-adk
mcp
starlette
uvicorn
flask
flask-cors
fastmcp